pymodules/solar-inverter/requirements.txt
pysolarmanv5
pyyaml
requests
configparser
paho-mqtt